Fixed Binary Search Subtraction Overflow Bug
The issue was that mid can be 0 and decrementing a usize 0 will cause a
overflow. Fixed by checking if mid is 0 and returning None if it is.
This is because if the right side of the bounds lies before the 0th
element, then the value is not in the list.